refactor: Moves modules to util
