+++ src/util/quack.ts
